Making performance claims its own file
